clk: qcom: move PAS clock group to its own compilation unit
PAS clock configuration is conditionally compiled and platform-specific;
keeping it in the same file as the core clock setup adds unnecessary
coupling and complicates maintenance.

No functional change.
