+++ core/drivers/clk/qcom/sub.mk
@@ -1,5 +1,5 @@
 global-incdirs-y += .
 
-srcs-y += clock-qcom.c
+srcs-y += clock-qcom.c clock-qcom-pas.c
 incdirs-y += .
-incdirs-y += platform/$(PLATFORM_FLAVOR)
+incdirs-$(CFG_QCOM_PAS_PTA) += platform/$(PLATFORM_FLAVOR)
